+++ examples/authentication/src/cortex/utils/types.ts
@@ -1,3 +1,0 @@
-import { services } from '../services/_services';
-
-export type Services = typeof services;