+++ examples/dog-api/src/cortex/dependencies/api/api.gateway.ts
@@ -1,9 +1,0 @@
-export type Breed = string;
-export type BreedVariant = string;
-export type ImageUri = string;
-export type ResponseStatus = 'success' | 'error';
-
-export interface ApiGateway {
-  getBreeds(): Promise<{ message: Record<Breed, BreedVariant[]>; status: ResponseStatus }>;
-  getBreedRandomImage(breed: Breed): Promise<{ message: ImageUri; status: ResponseStatus }>;
-}