+++ examples/dog-api/src/cortex/services/_services.ts
@@ -1,5 +1,0 @@
-import { DogService } from './dog.service';
-
-export const services = {
-  dog: DogService,
-};