+++ examples/todo/src/cortex/services/_services.ts
@@ -1,5 +1,0 @@
-import { TodoService } from './todo/todo.service';
-
-export const services = {
-  todo: TodoService,
-};