+++ packages/core/scripts/react/cortex/utils/hooks.ts
@@ -1,6 +1,0 @@
-// utils/hooks.ts
-
-import { createCortexHooks } from "@azot-dev/react-cortex";
-import { Services } from "./types";
-
-export const { useAppSelector, useLazyMethod, useMethod, useService, useStore, useAppState } = createCortexHooks<Services>();