+++ packages/core/src/react/core-context.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+import type { Core } from '../core'
+
+export const CoreContext = createContext<Core<any, any> | null>(null)