+++ packages/doc/src/pages/index.tsx
@@ -1,38 +1,6 @@
-import React from "react";
-import clsx from "clsx";
-import Link from "@docusaurus/Link";
-import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
-import Layout from "@theme/Layout";
-import HomepageFeatures from "@site/src/components/HomepageFeatures";
-
-import styles from "./index.module.css";
-
-function HomepageHeader() {
-  const { siteConfig } = useDocusaurusContext();
-  return (
-    <header className={clsx("hero hero--primary", styles.heroBanner)}>
-      <div className="container">
-        <img alt="logo" src={require("@site/static/img/logo_512.png").default} width="200px" height="200px" />
-        <h1 className="hero__title">{siteConfig.title}</h1>
-        <p className="hero__subtitle">{siteConfig.tagline}</p>
-        <div className={styles.buttons}>
-          <Link className="button button--secondary button--lg" to="/docs/setup">
-            Get Started
-          </Link>
-        </div>
-      </div>
-    </header>
-  );
-}
+import React, { JSX } from "react";
+import { Redirect } from "@docusaurus/router";
 
 export default function Home(): JSX.Element {
-  const { siteConfig } = useDocusaurusContext();
-  return (
-    <Layout title={`${siteConfig.title}`} description="THe missing brick of React to code in TDD and clean architecture">
-      <HomepageHeader />
-      <main>
-        <HomepageFeatures />
-      </main>
-    </Layout>
-  );
+  return <Redirect to="/docs/setup" />;
 }