+++ packages/example/src/cortex/services/_services.ts
@@ -1,5 +1,0 @@
-import { CounterService } from './counter.service';
-
-export const services = {
-  counter: CounterService,
-};