+++ packages/react/src/index.ts
@@ -1,3 +1,0 @@
-import { CortexProvider, createCortexHooks } from './provider';
-
-export { CortexProvider, createCortexHooks };